Moved most of update.update into related utilsm, since it is all reused by virtool ref add
