+++ virtool_cli/update/update_otu.py
@@ -5,9 +5,11 @@ import structlog
 
 from virtool_cli.utils.logging import DEFAULT_LOGGER, DEBUG_LOGGER
 from virtool_cli.utils.reference import get_otu_paths
-from virtool_cli.update.update import request_new_records, process_records, write_data
 from virtool_cli.utils.id_generator import get_unique_ids
+from virtool_cli.utils.format import process_records
+from virtool_cli.utils.storage import write_records
 from virtool_cli.catalog.helpers import search_by_otu_id
+from virtool_cli.update.update import request_new_records
 
 base_logger = structlog.get_logger()
 
@@ -78,4 +80,4 @@ async def update_otu(otu_path: Path, listing_path: Path, auto_evaluate: bool = F
     # List all isolate and sequence IDs presently in src
     unique_iso, unique_seq = await get_unique_ids(get_otu_paths(src_path))
 
-    await write_data(otu_path, otu_updates, unique_iso, unique_seq, logger=logger)
+    await write_records(otu_path, otu_updates, unique_iso, unique_seq, logger=logger)
