+++ virtool_cli/update/update_ref.py
@@ -6,8 +6,10 @@ import structlog
 from virtool_cli.utils.logging import DEFAULT_LOGGER, DEBUG_LOGGER
 from virtool_cli.utils.reference import get_otu_paths, search_otu_by_id, is_v1
 from virtool_cli.utils.id_generator import get_unique_ids
+from virtool_cli.utils.format import process_records
+from virtool_cli.utils.storage import write_records
+from virtool_cli.update.update import request_new_records
 from virtool_cli.catalog.helpers import filter_catalog
-from virtool_cli.update.update import request_new_records, process_records, write_data
 
 DEFAULT_INTERVAL = 0.001
 
@@ -227,7 +229,7 @@ async def writer_loop(
 
         otu_path = search_otu_by_id(otu_id, src_path)
 
-        await write_data(otu_path, sequence_data, unique_iso, unique_seq, logger)
+        await write_records(otu_path, sequence_data, unique_iso, unique_seq, logger)
 
         await asyncio.sleep(DEFAULT_INTERVAL)
         queue.task_done()
